Convert clickmenu into super class
